# Use span-based CopyTo/Async for ComputeHash/Async