Contributes to https://github.com/dotnet/corefx/issues/21472

cc: @stephentoub