@manandre 2019-11-30
I have an issue with a test:

```txt
System.Security.Cryptography.Hashing.Algorithms.Tests.InvalidUsageTests.InvalidHashCoreArgumentsFromStream [FAIL]
      Assert.Throws() Failure
      Expected: typeof(System.ArgumentException)
      Actual:   typeof(System.ArgumentOutOfRangeException): Specified argument was out of the range of valid values.
      ---- System.ArgumentOutOfRangeException : Specified argument was out of the range of valid values.
      Stack Trace:
        /_/src/System.Private.CoreLib/shared/System/IO/Stream.cs(226,0): at System.IO.Stream.WriteCallbackStream.Write(Byte[] buffer, Int32 offset, Int32 count)
        /_/src/System.Private.CoreLib/shared/System/IO/Stream.cs(153,0): at System.IO.Stream.CopyTo(Stream destination, Int32 bufferSize)
        /_/src/System.Private.CoreLib/shared/System/IO/Stream.cs(196,0): at System.IO.Stream.CopyTo(ReadOnlySpanAction`2 callback, Object state, Int32 bufferSize)
        /_/src/libraries/System.Security.Cryptography.Primitives/src/System/Security/Cryptography/HashAlgorithm.cs(102,0): at System.Security.Cryptography.HashAlgorithm.ComputeHash(Stream inputStream)
        /_/src/libraries/System.Security.Cryptography.Algorithms/tests/InvalidUsageTests.cs(31,0): at System.Security.Cryptography.Hashing.Algorithms.Tests.InvalidUsageTests.<>c__DisplayClass1_0.<InvalidHashCoreArgumentsFromStream>b__0()
```
[Helix logs](https://helix.dot.net/api/2019-06-17/jobs/58025783-e90b-4ef8-9fae-4fb246e9b738/workitems/System.Security.Cryptography.Algorithms.Tests/console)

I propose to change the expected exception from `ArgumentException` to `ArgumentOutOfRangeException` which is more adapted here.
As it is a behavior change at runtime level, I prefer that a member validates such a proposal. ❔