@manandre 2019-11-30
> Have you measured the performance of this, on streams that override the new CopyTo and those that don't? Can you share the test and results?

Performance test for MemoryStream based on `System.Security.Cryptography.Tests.Perf_Hashing` with:
```
[Benchmark]
public byte[] Sha() => _sha.ComputeHash(new {Custom}MemoryStream(_data));
```
and
```
public class CustomMemoryStream : MemoryStream
{
    public CustomMemoryStream(byte[] buffer) : base(buffer) { }
}
```
Results:

better: 6, geomean: 1.048
total diff: 6

No Slower results for the provided threshold = 0% and noise filter = 0.3ns.

| Faster                                                                                                                                  | base/diff | Base Median (ns) | Diff Median (ns) | Modality|
| --------------------------------------------------------------------------------------------------------------------------------------- | ---------:| ----------------:| ----------------:| --------:|
| System.Security.Cryptography.Tests.Perf_Hashing_Stream.Sha1(streamType: typeof(System.IO.MemoryStream)                                  |      1.10 |     273754450.00 |     249201400.00 |         |
| System.Security.Cryptography.Tests.Perf_Hashing_Stream.Sha512(streamType: typeof(System.IO.MemoryStream)                                |      1.06 |     377651400.00 |     355524700.00 |         |
| System.Security.Cryptography.Tests.Perf_Hashing_Stream.Sha384(streamType: typeof(System.IO.MemoryStream)                                |      1.06 |     378412200.00 |     357807450.00 |         |
| System.Security.Cryptography.Tests.Perf_Hashing_Stream.Sha256(streamType: typeof(System.IO.MemoryStream)                                |      1.04 |     616032400.00 |     591885000.00 |         |
| System.Security.Cryptography.Tests.Perf_Hashing_Stream.Sha384(streamType: typeof(System.Security.Cryptography.Tests.CustomMemoryStream) |      1.02 |     383589550.00 |     374913450.00 |         |
| System.Security.Cryptography.Tests.Perf_Hashing_Stream.Sha512(streamType: typeof(System.Security.Cryptography.Tests.CustomMemoryStream) |      1.01 |     377430800.00 |     374367200.00 |         |

Noticeable improvement 😎