@stephentoub 2019-12-01
> No slower

I expect this is because of https://github.com/dotnet/runtime/pull/391#discussion_r352361043.